More window manager hacking. FocusIn/FocusOut events.

+++ AK/AKString.h
@@ -65,9 +65,9 @@ public:
 
     bool is_null() const { return !m_impl; }
     bool is_empty() const { return length() == 0; }
-    unsigned length() const { return m_impl ? m_impl->length() : 0; }
+    size_t length() const { return m_impl ? m_impl->length() : 0; }
     const char* characters() const { return m_impl ? m_impl->characters() : nullptr; }
-    char operator[](unsigned i) const { ASSERT(m_impl); return (*m_impl)[i]; }
+    char operator[](size_t i) const { ASSERT(m_impl); return (*m_impl)[i]; }
 
     bool operator==(const String&) const;
     bool operator!=(const String& other) const { return !(*this == other); }
